+++ nas/README.md
@@ -1,0 +1,11 @@
+# Neural Architecture Search
+
+To help the researchers to do experiments on neural architecture search (NAS),
+we have implemented several baseline methods using the Auto-Keras framework.
+The implementations are easy since only the core part of the search algorithm is needed.
+All other parts of NAS (e.g. data structures for storing neural architectures, training of the neural networks)
+are done by the Auto-Keras framework.
+
+## Why implement NAS papers in Auto-Keras?
+
+## How to implement your own search?
